requirements.txt
Flask==2.1.3
Flask_SQLAlchemy==2.5.1
gunicorn==20.1.0
pandas==1.3.5
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
